requirements.txt
PyGLM
PyOgg
PyOpenAL
numpy
PyOpenGL
PyOpenGL-accelerate
colorama
glfw
Pillow
esper
freetype-py
pydub
pyglm-typing